drivers: elemu: move ELEMU_loadFw to SSSAPI middleware
This driver was using SSSAPI calls, whcih is wrong, as drivers should
not depend on middlewares. For backward compatibility, keep the API
name, but more it to the middleware codebase.

Additional change is to apply a reusable session ID for the loadFw call,
so we don't run into issues with a failing OPEN SESSION command when
trying to reopen a non-reusable (0) session.
