+++ mcux/mcux-sdk-ng/drivers/elemu/CMakeLists.txt
@@ -1,9 +1,9 @@
-# Copyright 2024-2025 NXP
+# Copyright 2024-2026 NXP
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
 if(CONFIG_MCUX_COMPONENT_driver.elemu)
-    mcux_component_version(2.1.3)
+    mcux_component_version(2.1.4)
 
     mcux_add_source(SOURCES fsl_elemu.c fsl_elemu.h)
