+++ mcux/mcux-sdk-ng/drivers/elemu/fsl_elemu.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2025 NXP
+ * Copyright 2019-2026 NXP
  * All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
@@ -18,12 +18,6 @@
 
 static status_t ELEMU_StartupWait(ELEMU_Type *mu);
 
-#if (defined(ELEMU_HAS_LOADABLE_FW) && ELEMU_HAS_LOADABLE_FW)
-#include "fsl_sss_mgmt.h"
-#include "fsl_sss_sscp.h"
-#include "fsl_sscp_mu.h"
-
-#endif
 /* ELEMU driver */
 
 /* undef BIT in case of redefinition
@@ -577,73 +571,3 @@ status_t ELEMU_LP_WakeupPathInit(ELEMU_Type *mu)
 
     return status;
 }
-
-#if (defined(ELEMU_HAS_LOADABLE_FW) && ELEMU_HAS_LOADABLE_FW)
-
-/* Tunnel type is EdgeLock Firmware upload */
-#define TUNNEL_TYPE_SB3_ELE_FW   0x22u
-#define SB3_BLOCK0_SIZE_IN_BYTES 0x60u
-
-/*!
- * brief Load firmware into ELE.
- *
- * This function uploads firmware to ELEMU using secure tunnel.
- *
- * param mu ELEMU peripheral base address
- * param image Pointer to firmware image data
- *
- * return kStatus_Success if successful, kStatus_Fail otherwise
- */
-status_t ELEMU_loadFw(ELEMU_Type *mu, uint32_t image[])
-{
-    uint32_t resultState          = 0u;
-    sss_sscp_tunnel_t tunnelCtx   = {0};
-    sss_sscp_session_t sssSession = {0};
-    sscp_context_t sscpContext    = {0};
-    status_t status               = kStatus_Fail;
-    do
-    {
-        if (sscp_mu_init(&(sscpContext), (ELEMU_Type *)mu) != kStatus_SSCP_Success)
-        {
-            break;
-        }
-        /* OPEN SESSION */
-        if (sss_sscp_open_session(&sssSession, 0u, kType_SSS_Ele200, &sscpContext) != kStatus_SSS_Success)
-        {
-            break;
-        }
-
-        /* OPEN TUNNEL */
-        if (sss_sscp_tunnel_context_init(&tunnelCtx, &sssSession, TUNNEL_TYPE_SB3_ELE_FW) != kStatus_SSS_Success)
-        {
-            break;
-        }
-
-        /* Pass loadable ELE FW */
-        if (sss_sscp_tunnel(&tunnelCtx, (uint8_t *)image, SB3_BLOCK0_SIZE_IN_BYTES, &resultState) !=
-            kStatus_SSS_Success)
-        {
-            break;
-        }
-
-        /* FREE TUNNEL CONTEXT */
-        if (sss_sscp_tunnel_context_free(&tunnelCtx) != kStatus_SSS_Success)
-        {
-            break;
-        }
-
-        /* CLOSE SESSION */
-        if (sss_sscp_close_session(&sssSession) != kStatus_SSS_Success)
-        {
-            break;
-        }
-
-        /* If all steps before passes without break, then consider it as success*/
-        status = kStatus_Success;
-
-    } while (false);
-
-    return status;
-}
-
-#endif /* ELEMU_HAS_LOADABLE_FW */
