+++ mcux/mcux-sdk-ng/drivers/elemu/fsl_elemu.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2025 NXP
+ * Copyright 2019-2026 NXP
  * All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
@@ -227,20 +227,6 @@ void ELEMU_mu_init(ELEMU_Type *mu);
  * @return kStatus_Success if successful, kStatus_Fail otherwise
  */
 status_t ELEMU_LP_WakeupPathInit(ELEMU_Type *mu);
-
-#if (defined(ELEMU_HAS_LOADABLE_FW) && ELEMU_HAS_LOADABLE_FW)
-/*!
- * @brief Load firmware into ELE.
- *
- * This function uploads firmware to ELEMU using secure tunnel.
- *
- * @param mu ELEMU peripheral base address
- * @param image Pointer to firmware image data
- *
- * @return kStatus_Success if successful, kStatus_Fail otherwise
- */
-status_t ELEMU_loadFw(ELEMU_Type *mu, uint32_t image[]);
-#endif /* ELEMU_HAS_LOADABLE_FW */
 
 /** @}*/
